x86_vcpu

Enum VmxExitReason

#[repr(u32)]
pub enum VmxExitReason {

    EXCEPTION_NMI = 0,
    EXTERNAL_INTERRUPT = 1,
    TRIPLE_FAULT = 2,
    INIT = 3,
    SIPI = 4,
    SMI = 5,
    OTHER_SMI = 6,
    INTERRUPT_WINDOW = 7,
    NMI_WINDOW = 8,
    TASK_SWITCH = 9,
    CPUID = 10,
    GETSEC = 11,
    HLT = 12,
    INVD = 13,
    INVLPG = 14,
    RDPMC = 15,
    RDTSC = 16,
    RSM = 17,
    VMCALL = 18,
    VMCLEAR = 19,
    VMLAUNCH = 20,
    VMPTRLD = 21,
    VMPTRST = 22,
    VMREAD = 23,
    VMRESUME = 24,
    VMWRITE = 25,
    VMOFF = 26,
    VMON = 27,
    CR_ACCESS = 28,
    DR_ACCESS = 29,
    IO_INSTRUCTION = 30,
    MSR_READ = 31,
    MSR_WRITE = 32,
    INVALID_GUEST_STATE = 33,
    MSR_LOAD_FAIL = 34,
    MWAIT_INSTRUCTION = 36,
    MONITOR_TRAP_FLAG = 37,
    MONITOR_INSTRUCTION = 39,
    PAUSE_INSTRUCTION = 40,
    MCE_DURING_VMENTRY = 41,
    TPR_BELOW_THRESHOLD = 43,
    APIC_ACCESS = 44,
    VIRTUALIZED_EOI = 45,
    GDTR_IDTR = 46,
    LDTR_TR = 47,
    EPT_VIOLATION = 48,
    EPT_MISCONFIG = 49,
    INVEPT = 50,
    RDTSCP = 51,
    PREEMPTION_TIMER = 52,
    INVVPID = 53,
    WBINVD = 54,
    XSETBV = 55,
    APIC_WRITE = 56,
    RDRAND = 57,
    INVPCID = 58,
    VMFUNC = 59,
    ENCLS = 60,
    RDSEED = 61,
    PML_FULL = 62,
    XSAVES = 63,
    XRSTORS = 64,
    PCONFIG = 65,
    SPP_EVENT = 66,
    UMWAIT = 67,
    TPAUSE = 68,
    LOADIWKEY = 69,

}

VMX basic exit reasons, as defined in the Intel Software Developerโs Manual (SDM) Vol. 3D, Appendix C.

This enum represents the various reasons why a VM exit might occur during the execution of a virtual machine in VMX (Virtual Machine Extensions) mode. Each variant corresponds to a specific exit reason that can be identified and handled by the hypervisor.

Variants

EXCEPTION_NMI = 0

Exception or non-maskable interrupt (NMI) occurred.

EXTERNAL_INTERRUPT = 1

An external interrupt was received.

TRIPLE_FAULT = 2

A triple fault occurred.

INIT = 3

INIT signal was received.

SIPI = 4

Startup IPI (SIPI) was received.

SMI = 5

System Management Interrupt (SMI) was received.

OTHER_SMI = 6

Other SMI was received.

INTERRUPT_WINDOW = 7

An interrupt window was open.

NMI_WINDOW = 8

An NMI window was open.

TASK_SWITCH = 9

A task switch occurred.

CPUID = 10

CPUID instruction was executed.

GETSEC = 11

GETSEC instruction was executed.

HLT = 12

HLT instruction was executed.

INVD = 13

INVD instruction was executed.

INVLPG = 14

INVLPG instruction was executed.

RDPMC = 15

RDPMC instruction was executed.

RDTSC = 16

RDTSC instruction was executed.

RSM = 17

RSM instruction was executed in SMM.

VMCALL = 18

VMCALL instruction was executed.

VMCLEAR = 19

VMCLEAR instruction was executed.

VMLAUNCH = 20

VMLAUNCH instruction was executed.

VMPTRLD = 21

VMPTRLD instruction was executed.

VMPTRST = 22

VMPTRST instruction was executed.

VMREAD = 23

VMREAD instruction was executed.

VMRESUME = 24

VMRESUME instruction was executed.

VMWRITE = 25

VMWRITE instruction was executed.

VMOFF = 26

VMOFF instruction was executed.

VMON = 27

VMON instruction was executed.

CR_ACCESS = 28

Control Register (CR) access.

DR_ACCESS = 29

Debug Register (DR) access.

IO_INSTRUCTION = 30

I/O instruction was executed.

MSR_READ = 31

Model-Specific Register (MSR) read.

MSR_WRITE = 32

Model-Specific Register (MSR) write.

INVALID_GUEST_STATE = 33

Guest state is invalid.

MSR_LOAD_FAIL = 34

MSR load failed.

MWAIT_INSTRUCTION = 36

MWAIT instruction was executed.

MONITOR_TRAP_FLAG = 37

Monitor trap flag triggered.

MONITOR_INSTRUCTION = 39

MONITOR instruction was executed.

PAUSE_INSTRUCTION = 40

PAUSE instruction was executed.

MCE_DURING_VMENTRY = 41

Machine Check Exception (MCE) occurred during VM entry.

TPR_BELOW_THRESHOLD = 43

Task Priority Register (TPR) below threshold.

APIC_ACCESS = 44

Access to Advanced Programmable Interrupt Controller (APIC).

VIRTUALIZED_EOI = 45

Virtualized End Of Interrupt (EOI) was executed.

GDTR_IDTR = 46

Access to Global Descriptor Table Register (GDTR) or Interrupt Descriptor Table Register (IDTR).

LDTR_TR = 47

Access to Local Descriptor Table Register (LDTR) or Task Register (TR).

EPT_VIOLATION = 48

Extended Page Table (EPT) violation occurred.

EPT_MISCONFIG = 49

Extended Page Table (EPT) misconfiguration occurred.

INVEPT = 50

INVEPT instruction was executed.

RDTSCP = 51

RDTSCP instruction was executed.

PREEMPTION_TIMER = 52

Preemption timer expired.

INVVPID = 53

INVVPID instruction was executed.

WBINVD = 54

WBINVD instruction was executed.

XSETBV = 55

XSETBV instruction was executed.

APIC_WRITE = 56

APIC write occurred.

RDRAND = 57

RDRAND instruction was executed.

INVPCID = 58

INVPCID instruction was executed.

VMFUNC = 59

VMFUNC was executed.

ENCLS = 60

ENCLS instruction was executed.

RDSEED = 61

RDSEED instruction was executed.

PML_FULL = 62

Page modification log (PML) became full.

XSAVES = 63

XSAVES instruction was executed.

XRSTORS = 64

XRSTORS instruction was executed.

PCONFIG = 65

PCONFIG instruction was executed.

SPP_EVENT = 66

SPP event occurred.

UMWAIT = 67

UMWAIT instruction was executed.

TPAUSE = 68

TPAUSE instruction was executed.

LOADIWKEY = 69

LOADIWKEY instruction was executed.