x86_vcpu
Struct GeneralRegisters

#[repr(C)]
pub struct GeneralRegisters {

    pub rax: u64,
    pub rcx: u64,
    pub rdx: u64,
    pub rbx: u64,
    pub rbp: u64,
    pub rsi: u64,
    pub rdi: u64,
    pub r8: u64,
    pub r9: u64,
    pub r10: u64,
    pub r11: u64,
    pub r12: u64,
    pub r13: u64,
    pub r14: u64,
    pub r15: u64,
    /* private fields */

}

General-purpose registers for the 64-bit x86 architecture.

This structure holds the values of the general-purpose registers used in 64-bit x86 systems, allowing for easy manipulation and storage of register states.

Fields
rax: u64

The RAX register, typically used for return values in functions.

rcx: u64

The RCX register, often used as a counter in loops.

rdx: u64

The RDX register, commonly used for I/O operations.

rbx: u64

The RBX register, usually used as a base pointer or to store values across function calls.

rbp: u64

The RBP register, often used as a frame pointer in function calls.

rsi: u64

The RSI register, often used as a source index in string operations.

rdi: u64

The RDI register, often used as a destination index in string operations.

r8: u64

The R8 register, an additional general-purpose register available in 64-bit mode.

r9: u64

The R9 register, an additional general-purpose register available in 64-bit mode.

r10: u64

The R10 register, an additional general-purpose register available in 64-bit mode.

r11: u64

The R11 register, an additional general-purpose register available in 64-bit mode.

r12: u64

The R12 register, an additional general-purpose register available in 64-bit mode.

r13: u64

The R13 register, an additional general-purpose register available in 64-bit mode.

r14: u64

The R14 register, an additional general-purpose register available in 64-bit mode.

r15: u64

The R15 register, an additional general-purpose register available in 64-bit mode.

Implementations

impl GeneralRegisters

pub fn get_reg_of_index(&self, index: u8) -> u64

Returns the value of the general-purpose register corresponding to the given index.

The mapping of indices to registers is as follows:

0: rax

1: rcx

2: rdx

3: rbx

5: rbp

6: rsi

7: rdi

8: r8

9: r9

10: r10

11: r11

12: r12

13: r13

14: r14

15: r15

Panics

This function will panic if the provided index is out of the range [0, 15] or if the index corresponds to an unused register (rsp at index 4).

Arguments

index - A u8 value representing the index of the register.

Returns

u64 - The value of the corresponding general-purpose register.

pub fn set_reg_of_index(&mut self, index: u8, value: u64)

Sets the value of the general-purpose register corresponding to the given index.

The mapping of indices to registers is as follows:

0: rax

1: rcx

2: rdx

3: rbx

5: rbp

6: rsi

7: rdi

8: r8

9: r9

10: r10

11: r11

12: r12

13: r13

14: r14

15: r15

Panics

This function will panic if the provided index is out of the range [0, 15] or if the index corresponds to an unused register (rsp at index 4).

Arguments

index - A u8 value representing the index of the register.

Returns

u64 - The value of the corresponding general-purpose register.

pub fn get_edx_eax(&self) -> u64

Returns the value of the edx:eax register pair.

impl GeneralRegisters

pub fn eax(&self) -> u32

Returns the value of the “eax” register, which is the bits 0(inc.) to 32(exc.) of the “rax” register.

pub fn set_eax(&mut self, value: u32)

Sets the value of the “eax” register, which is the bits 0(inc.) to 32(exc.) of the “rax” register.

Whether the other bits of the “rax” register should be cleared follows the rules of the x86-64 architecture.

pub fn ecx(&self) -> u32

Returns the value of the “ecx” register, which is the bits 0(inc.) to 32(exc.) of the “rcx” register.

pub fn set_ecx(&mut self, value: u32)

Sets the value of the “ecx” register, which is the bits 0(inc.) to 32(exc.) of the “rcx” register.

Whether the other bits of the “rcx” register should be cleared follows the rules of the x86-64 architecture.

pub fn edx(&self) -> u32

Returns the value of the “edx” register, which is the bits 0(inc.) to 32(exc.) of the “rdx” register.

pub fn set_edx(&mut self, value: u32)

Sets the value of the “edx” register, which is the bits 0(inc.) to 32(exc.) of the “rdx” register.

Whether the other bits of the “rdx” register should be cleared follows the rules of the x86-64 architecture.

pub fn ebx(&self) -> u32

Returns the value of the “ebx” register, which is the bits 0(inc.) to 32(exc.) of the “rbx” register.

pub fn set_ebx(&mut self, value: u32)

Sets the value of the “ebx” register, which is the bits 0(inc.) to 32(exc.) of the “rbx” register.

Whether the other bits of the “rbx” register should be cleared follows the rules of the x86-64 architecture.

pub fn ebp(&self) -> u32

Returns the value of the “ebp” register, which is the bits 0(inc.) to 32(exc.) of the “rbp” register.

pub fn set_ebp(&mut self, value: u32)

Sets the value of the “ebp” register, which is the bits 0(inc.) to 32(exc.) of the “rbp” register.

Whether the other bits of the “rbp” register should be cleared follows the rules of the x86-64 architecture.

pub fn esi(&self) -> u32

Returns the value of the “esi” register, which is the bits 0(inc.) to 32(exc.) of the “rsi” register.

pub fn set_esi(&mut self, value: u32)

Sets the value of the “esi” register, which is the bits 0(inc.) to 32(exc.) of the “rsi” register.

Whether the other bits of the “rsi” register should be cleared follows the rules of the x86-64 architecture.

pub fn edi(&self) -> u32

Returns the value of the “edi” register, which is the bits 0(inc.) to 32(exc.) of the “rdi” register.

pub fn set_edi(&mut self, value: u32)

Sets the value of the “edi” register, which is the bits 0(inc.) to 32(exc.) of the “rdi” register.

Whether the other bits of the “rdi” register should be cleared follows the rules of the x86-64 architecture.

pub fn r8d(&self) -> u32

Returns the value of the “r8d” register, which is the bits 0(inc.) to 32(exc.) of the “r8” register.

pub fn set_r8d(&mut self, value: u32)

Sets the value of the “r8d” register, which is the bits 0(inc.) to 32(exc.) of the “r8” register.

Whether the other bits of the “r8” register should be cleared follows the rules of the x86-64 architecture.

pub fn r9d(&self) -> u32

Returns the value of the “r9d” register, which is the bits 0(inc.) to 32(exc.) of the “r9” register.

pub fn set_r9d(&mut self, value: u32)

Sets the value of the “r9d” register, which is the bits 0(inc.) to 32(exc.) of the “r9” register.

Whether the other bits of the “r9” register should be cleared follows the rules of the x86-64 architecture.

pub fn r10d(&self) -> u32

Returns the value of the “r10d” register, which is the bits 0(inc.) to 32(exc.) of the “r10” register.

pub fn set_r10d(&mut self, value: u32)

Sets the value of the “r10d” register, which is the bits 0(inc.) to 32(exc.) of the “r10” register.

Whether the other bits of the “r10” register should be cleared follows the rules of the x86-64 architecture.

pub fn r11d(&self) -> u32

Returns the value of the “r11d” register, which is the bits 0(inc.) to 32(exc.) of the “r11” register.

pub fn set_r11d(&mut self, value: u32)

Sets the value of the “r11d” register, which is the bits 0(inc.) to 32(exc.) of the “r11” register.

Whether the other bits of the “r11” register should be cleared follows the rules of the x86-64 architecture.

pub fn r12d(&self) -> u32

Returns the value of the “r12d” register, which is the bits 0(inc.) to 32(exc.) of the “r12” register.

pub fn set_r12d(&mut self, value: u32)

Sets the value of the “r12d” register, which is the bits 0(inc.) to 32(exc.) of the “r12” register.

Whether the other bits of the “r12” register should be cleared follows the rules of the x86-64 architecture.

pub fn r13d(&self) -> u32

Returns the value of the “r13d” register, which is the bits 0(inc.) to 32(exc.) of the “r13” register.

pub fn set_r13d(&mut self, value: u32)

Sets the value of the “r13d” register, which is the bits 0(inc.) to 32(exc.) of the “r13” register.

Whether the other bits of the “r13” register should be cleared follows the rules of the x86-64 architecture.

pub fn r14d(&self) -> u32

Returns the value of the “r14d” register, which is the bits 0(inc.) to 32(exc.) of the “r14” register.

pub fn set_r14d(&mut self, value: u32)

Sets the value of the “r14d” register, which is the bits 0(inc.) to 32(exc.) of the “r14” register.

Whether the other bits of the “r14” register should be cleared follows the rules of the x86-64 architecture.

pub fn r15d(&self) -> u32

Returns the value of the “r15d” register, which is the bits 0(inc.) to 32(exc.) of the “r15” register.

pub fn set_r15d(&mut self, value: u32)

Sets the value of the “r15d” register, which is the bits 0(inc.) to 32(exc.) of the “r15” register.

Whether the other bits of the “r15” register should be cleared follows the rules of the x86-64 architecture.

pub fn ax(&self) -> u16

Returns the value of the “ax” register, which is the bits 0(inc.) to 16(exc.) of the “rax” register.

pub fn set_ax(&mut self, value: u16)

Sets the value of the “ax” register, which is the bits 0(inc.) to 16(exc.) of the “rax” register.

Whether the other bits of the “rax” register should be cleared follows the rules of the x86-64 architecture.

pub fn cx(&self) -> u16

Returns the value of the “cx” register, which is the bits 0(inc.) to 16(exc.) of the “rcx” register.

pub fn set_cx(&mut self, value: u16)

Sets the value of the “cx” register, which is the bits 0(inc.) to 16(exc.) of the “rcx” register.

Whether the other bits of the “rcx” register should be cleared follows the rules of the x86-64 architecture.

pub fn dx(&self) -> u16

Returns the value of the “dx” register, which is the bits 0(inc.) to 16(exc.) of the “rdx” register.

pub fn set_dx(&mut self, value: u16)

Sets the value of the “dx” register, which is the bits 0(inc.) to 16(exc.) of the “rdx” register.

Whether the other bits of the “rdx” register should be cleared follows the rules of the x86-64 architecture.

pub fn bx(&self) -> u16

Returns the value of the “bx” register, which is the bits 0(inc.) to 16(exc.) of the “rbx” register.

pub fn set_bx(&mut self, value: u16)

Sets the value of the “bx” register, which is the bits 0(inc.) to 16(exc.) of the “rbx” register.

Whether the other bits of the “rbx” register should be cleared follows the rules of the x86-64 architecture.

pub fn bp(&self) -> u16

Returns the value of the “bp” register, which is the bits 0(inc.) to 16(exc.) of the “rbp” register.

pub fn set_bp(&mut self, value: u16)

Sets the value of the “bp” register, which is the bits 0(inc.) to 16(exc.) of the “rbp” register.

Whether the other bits of the “rbp” register should be cleared follows the rules of the x86-64 architecture.

pub fn si(&self) -> u16

Returns the value of the “si” register, which is the bits 0(inc.) to 16(exc.) of the “rsi” register.

pub fn set_si(&mut self, value: u16)

Sets the value of the “si” register, which is the bits 0(inc.) to 16(exc.) of the “rsi” register.

Whether the other bits of the “rsi” register should be cleared follows the rules of the x86-64 architecture.

pub fn di(&self) -> u16

Returns the value of the “di” register, which is the bits 0(inc.) to 16(exc.) of the “rdi” register.

pub fn set_di(&mut self, value: u16)

Sets the value of the “di” register, which is the bits 0(inc.) to 16(exc.) of the “rdi” register.

Whether the other bits of the “rdi” register should be cleared follows the rules of the x86-64 architecture.

pub fn r8w(&self) -> u16

Returns the value of the “r8w” register, which is the bits 0(inc.) to 16(exc.) of the “r8” register.

pub fn set_r8w(&mut self, value: u16)

Sets the value of the “r8w” register, which is the bits 0(inc.) to 16(exc.) of the “r8” register.

Whether the other bits of the “r8” register should be cleared follows the rules of the x86-64 architecture.

pub fn r9w(&self) -> u16

Returns the value of the “r9w” register, which is the bits 0(inc.) to 16(exc.) of the “r9” register.

pub fn set_r9w(&mut self, value: u16)

Sets the value of the “r9w” register, which is the bits 0(inc.) to 16(exc.) of the “r9” register.

Whether the other bits of the “r9” register should be cleared follows the rules of the x86-64 architecture.

pub fn r10w(&self) -> u16

Returns the value of the “r10w” register, which is the bits 0(inc.) to 16(exc.) of the “r10” register.

pub fn set_r10w(&mut self, value: u16)

Sets the value of the “r10w” register, which is the bits 0(inc.) to 16(exc.) of the “r10” register.

Whether the other bits of the “r10” register should be cleared follows the rules of the x86-64 architecture.

pub fn r11w(&self) -> u16

Returns the value of the “r11w” register, which is the bits 0(inc.) to 16(exc.) of the “r11” register.

pub fn set_r11w(&mut self, value: u16)

Sets the value of the “r11w” register, which is the bits 0(inc.) to 16(exc.) of the “r11” register.

Whether the other bits of the “r11” register should be cleared follows the rules of the x86-64 architecture.

pub fn r12w(&self) -> u16

Returns the value of the “r12w” register, which is the bits 0(inc.) to 16(exc.) of the “r12” register.

pub fn set_r12w(&mut self, value: u16)

Sets the value of the “r12w” register, which is the bits 0(inc.) to 16(exc.) of the “r12” register.

Whether the other bits of the “r12” register should be cleared follows the rules of the x86-64 architecture.

pub fn r13w(&self) -> u16

Returns the value of the “r13w” register, which is the bits 0(inc.) to 16(exc.) of the “r13” register.

pub fn set_r13w(&mut self, value: u16)

Sets the value of the “r13w” register, which is the bits 0(inc.) to 16(exc.) of the “r13” register.

Whether the other bits of the “r13” register should be cleared follows the rules of the x86-64 architecture.

pub fn r14w(&self) -> u16

Returns the value of the “r14w” register, which is the bits 0(inc.) to 16(exc.) of the “r14” register.

pub fn set_r14w(&mut self, value: u16)

Sets the value of the “r14w” register, which is the bits 0(inc.) to 16(exc.) of the “r14” register.

Whether the other bits of the “r14” register should be cleared follows the rules of the x86-64 architecture.

pub fn r15w(&self) -> u16

Returns the value of the “r15w” register, which is the bits 0(inc.) to 16(exc.) of the “r15” register.

pub fn set_r15w(&mut self, value: u16)

Sets the value of the “r15w” register, which is the bits 0(inc.) to 16(exc.) of the “r15” register.

Whether the other bits of the “r15” register should be cleared follows the rules of the x86-64 architecture.

pub fn al(&self) -> u8

Returns the value of the “al” register, which is the bits 0(inc.) to 8(exc.) of the “rax” register.

pub fn set_al(&mut self, value: u8)

Sets the value of the “al” register, which is the bits 0(inc.) to 8(exc.) of the “rax” register.

Whether the other bits of the “rax” register should be cleared follows the rules of the x86-64 architecture.

pub fn cl(&self) -> u8

Returns the value of the “cl” register, which is the bits 0(inc.) to 8(exc.) of the “rcx” register.

pub fn set_cl(&mut self, value: u8)

Sets the value of the “cl” register, which is the bits 0(inc.) to 8(exc.) of the “rcx” register.

Whether the other bits of the “rcx” register should be cleared follows the rules of the x86-64 architecture.

pub fn dl(&self) -> u8

Returns the value of the “dl” register, which is the bits 0(inc.) to 8(exc.) of the “rdx” register.

pub fn set_dl(&mut self, value: u8)

Sets the value of the “dl” register, which is the bits 0(inc.) to 8(exc.) of the “rdx” register.

Whether the other bits of the “rdx” register should be cleared follows the rules of the x86-64 architecture.

pub fn bl(&self) -> u8

Returns the value of the “bl” register, which is the bits 0(inc.) to 8(exc.) of the “rbx” register.

pub fn set_bl(&mut self, value: u8)

Sets the value of the “bl” register, which is the bits 0(inc.) to 8(exc.) of the “rbx” register.

Whether the other bits of the “rbx” register should be cleared follows the rules of the x86-64 architecture.

pub fn bpl(&self) -> u8

Returns the value of the “bpl” register, which is the bits 0(inc.) to 8(exc.) of the “rbp” register.

pub fn set_bpl(&mut self, value: u8)

Sets the value of the “bpl” register, which is the bits 0(inc.) to 8(exc.) of the “rbp” register.

Whether the other bits of the “rbp” register should be cleared follows the rules of the x86-64 architecture.

pub fn sil(&self) -> u8

Returns the value of the “sil” register, which is the bits 0(inc.) to 8(exc.) of the “rsi” register.

pub fn set_sil(&mut self, value: u8)

Sets the value of the “sil” register, which is the bits 0(inc.) to 8(exc.) of the “rsi” register.

Whether the other bits of the “rsi” register should be cleared follows the rules of the x86-64 architecture.

pub fn dil(&self) -> u8

Returns the value of the “dil” register, which is the bits 0(inc.) to 8(exc.) of the “rdi” register.

pub fn set_dil(&mut self, value: u8)

Sets the value of the “dil” register, which is the bits 0(inc.) to 8(exc.) of the “rdi” register.

Whether the other bits of the “rdi” register should be cleared follows the rules of the x86-64 architecture.

pub fn r8b(&self) -> u8

Returns the value of the “r8b” register, which is the bits 0(inc.) to 8(exc.) of the “r8” register.

pub fn set_r8b(&mut self, value: u8)

Sets the value of the “r8b” register, which is the bits 0(inc.) to 8(exc.) of the “r8” register.

Whether the other bits of the “r8” register should be cleared follows the rules of the x86-64 architecture.

pub fn r9b(&self) -> u8

Returns the value of the “r9b” register, which is the bits 0(inc.) to 8(exc.) of the “r9” register.

pub fn set_r9b(&mut self, value: u8)

Sets the value of the “r9b” register, which is the bits 0(inc.) to 8(exc.) of the “r9” register.

Whether the other bits of the “r9” register should be cleared follows the rules of the x86-64 architecture.

pub fn r10b(&self) -> u8

Returns the value of the “r10b” register, which is the bits 0(inc.) to 8(exc.) of the “r10” register.

pub fn set_r10b(&mut self, value: u8)

Sets the value of the “r10b” register, which is the bits 0(inc.) to 8(exc.) of the “r10” register.

Whether the other bits of the “r10” register should be cleared follows the rules of the x86-64 architecture.

pub fn r11b(&self) -> u8

Returns the value of the “r11b” register, which is the bits 0(inc.) to 8(exc.) of the “r11” register.

pub fn set_r11b(&mut self, value: u8)

Sets the value of the “r11b” register, which is the bits 0(inc.) to 8(exc.) of the “r11” register.

Whether the other bits of the “r11” register should be cleared follows the rules of the x86-64 architecture.

pub fn r12b(&self) -> u8

Returns the value of the “r12b” register, which is the bits 0(inc.) to 8(exc.) of the “r12” register.

pub fn set_r12b(&mut self, value: u8)

Sets the value of the “r12b” register, which is the bits 0(inc.) to 8(exc.) of the “r12” register.

Whether the other bits of the “r12” register should be cleared follows the rules of the x86-64 architecture.

pub fn r13b(&self) -> u8

Returns the value of the “r13b” register, which is the bits 0(inc.) to 8(exc.) of the “r13” register.

pub fn set_r13b(&mut self, value: u8)

Sets the value of the “r13b” register, which is the bits 0(inc.) to 8(exc.) of the “r13” register.

Whether the other bits of the “r13” register should be cleared follows the rules of the x86-64 architecture.

pub fn r14b(&self) -> u8

Returns the value of the “r14b” register, which is the bits 0(inc.) to 8(exc.) of the “r14” register.

pub fn set_r14b(&mut self, value: u8)

Sets the value of the “r14b” register, which is the bits 0(inc.) to 8(exc.) of the “r14” register.

Whether the other bits of the “r14” register should be cleared follows the rules of the x86-64 architecture.

pub fn r15b(&self) -> u8

Returns the value of the “r15b” register, which is the bits 0(inc.) to 8(exc.) of the “r15” register.

pub fn set_r15b(&mut self, value: u8)

Sets the value of the “r15b” register, which is the bits 0(inc.) to 8(exc.) of the “r15” register.

Whether the other bits of the “r15” register should be cleared follows the rules of the x86-64 architecture.

pub fn ah(&self) -> u8

Returns the value of the “ah” register, which is the bits 8(inc.) to 16(exc.) of the “rax” register.

pub fn set_ah(&mut self, value: u8)

Sets the value of the “ah” register, which is the bits 8(inc.) to 16(exc.) of the “rax” register.

Whether the other bits of the “rax” register should be cleared follows the rules of the x86-64 architecture.

pub fn ch(&self) -> u8

Returns the value of the “ch” register, which is the bits 8(inc.) to 16(exc.) of the “rcx” register.

pub fn set_ch(&mut self, value: u8)

Sets the value of the “ch” register, which is the bits 8(inc.) to 16(exc.) of the “rcx” register.

Whether the other bits of the “rcx” register should be cleared follows the rules of the x86-64 architecture.

pub fn dh(&self) -> u8

Returns the value of the “dh” register, which is the bits 8(inc.) to 16(exc.) of the “rdx” register.

pub fn set_dh(&mut self, value: u8)

Sets the value of the “dh” register, which is the bits 8(inc.) to 16(exc.) of the “rdx” register.

Whether the other bits of the “rdx” register should be cleared follows the rules of the x86-64 architecture.

pub fn bh(&self) -> u8

Returns the value of the “bh” register, which is the bits 8(inc.) to 16(exc.) of the “rbx” register.

pub fn set_bh(&mut self, value: u8)

Sets the value of the “bh” register, which is the bits 8(inc.) to 16(exc.) of the “rbx” register.

Whether the other bits of the “rbx” register should be cleared follows the rules of the x86-64 architecture.

impl GeneralRegisters

pub const REGISTER_NAMES: [&'static str; 16]

The names of the general-purpose registers in 64-bit x86 architecture.

We follow the order of registers opcode encoding.

pub const fn register_name(index: u8) -> &'static str

Returns the name of the register corresponding to the given index.