x86_vcpu
Struct VmxArchVCpu

#[repr(C)]
pub struct VmxArchVCpu<H: AxVCpuHal> { /* private fields */ }

A virtual CPU within a guest.

Implementations

impl<H: AxVCpuHal> VmxVcpu<H>

pub fn new(vm_id: VMId, vcpu_id: VCpuId) -> AxResult<Self>

Create a new VmxVcpu.

pub fn setup( &mut self, ept_root: HostPhysAddr, entry: GuestPhysAddr, ) -> AxResult

Set the new VmxVcpu context from guest OS.

pub fn bind_to_current_processor(&self) -> AxResult

Bind this VmxVcpu to current logical processor.

pub fn unbind_from_current_processor(&self) -> AxResult

Unbind this VmxVcpu from current logical processor.

pub fn get_cpu_mode(&self) -> VmCpuMode

Get CPU mode of the guest.

pub fn inner_run(&mut self) -> Option<VmxExitInfo>

Run the guest. It returns when a vm-exit happens and returns the vm-exit if it cannot be handled by this VmxVcpu itself.

pub fn exit_info(&self) -> AxResult<VmxExitInfo>

Basic information about VM exits.

pub fn raw_interrupt_exit_info(&self) -> AxResult<u32>

Raw information for VM Exits Due to Vectored Events, See SDM 25.9.2

pub fn interrupt_exit_info(&self) -> AxResult<VmxInterruptInfo>

Information for VM exits due to external interrupts.

pub fn io_exit_info(&self) -> AxResult<VmxIoExitInfo>

Information for VM exits due to I/O instructions.

pub fn nested_page_fault_info(&self) -> AxResult<NestedPageFaultInfo>

Information for VM exits due to nested page table faults (EPT violation).

pub fn apic_access_exit_info(&self) -> AxResult<ApicAccessExitInfo>

Information for VM exits due to APIC access.

pub fn regs(&self) -> &GeneralRegisters

Guest general-purpose registers.

pub fn regs_mut(&mut self) -> &mut GeneralRegisters

Mutable reference of guest general-purpose registers.

pub fn stack_pointer(&self) -> usize

Guest stack pointer. (RSP)

pub fn set_stack_pointer(&mut self, rsp: usize)

Set guest stack pointer. (RSP)

pub fn gla2gva(&self, guest_rip: GuestVirtAddr) -> GuestVirtAddr

Translate guest virtual addr to linear addr

pub fn get_ptw_info(&self) -> GuestPageWalkInfo

Get Translate guest page table info

pub fn rip(&self) -> usize

Guest rip. (RIP)

pub fn cs(&self) -> u16

Guest cs. (cs)

pub fn advance_rip(&mut self, instr_len: u8) -> AxResult

Advance guest RIP by instr_len bytes.

pub fn queue_event(&mut self, vector: u8, err_code: Option<u32>)

Add a virtual interrupt or exception to the pending events list, and try to inject it before later VM entries.

pub fn set_interrupt_window(&mut self, enable: bool) -> AxResult

If enable, a VM exit occurs at the beginning of any instruction if RFLAGS.IF = 1 and there are no other blocking of interrupts. (see SDM, Vol. 3C, Section 24.4.2)

pub fn set_io_intercept_of_range( &mut self, port_base: u32, count: u32, intercept: bool, )

Set I/O intercept by modifying I/O bitmap.

pub fn set_msr_intercept_of_range(&mut self, msr: u32, intercept: bool)

Set msr intercept by modifying msr bitmap. Todo: distinguish read and write.