Trait x86_vcpu::PhysFrameIf

pub trait PhysFrameIf {
    // Required methods
    fn alloc_frame() -> Option<PhysAddr>;
    fn dealloc_frame(paddr: PhysAddr);
    fn phys_to_virt(paddr: PhysAddr) -> VirtAddr;
}

Low-level resource interfaces that must be implemented by the crate user.

Required Methods

fn alloc_frame() -> Option<PhysAddr>

Request to allocate a 4K-sized physical frame.

fn dealloc_frame(paddr: PhysAddr)

Request to free a allocated physical frame.

fn phys_to_virt(paddr: PhysAddr) -> VirtAddr

Returns a virtual address that maps to the given physical address.

Used to access the physical memory directly in PhysFrame implementation through as_mut_ptr().

Object Safety

This trait is not object safe.