Crate x86_vcpu

x86_vcpu

Definition of the vCPU structure and virtualization-related interface support for x86_64 architecture.

The crate user must implement the PhysFrameIf trait using crate_interface::impl_interface to provide the low-level implementantion of the allocation and dealloction of PhysFrame, relevant implementation can refer to ArceOS.

Example

use x86_vcpu::PhysFrameIf;

struct PhysFrameIfImpl;

#[crate_interface::impl_interface]
impl axvm::PhysFrameIf for PhysFrameIfImpl {
    fn alloc_frame() -> Option<PhysAddr> {
        // Your implementation here
    }
    fn dealloc_frame(paddr: PhysAddr) {
        // Your implementation here
    }
    fn phys_to_virt(paddr: PhysAddr) -> VirtAddr {
        // Your implementation here
    }
}

Structs

GeneralRegisters

General-purpose registers for the 64-bit x86 architecture.

GuestPageWalkInfo

The information of guest page walk.

VmxArchPerCpuState

Represents the per-CPU state for Virtual Machine Extensions (VMX).

VmxArchVCpu

A virtual CPU within a guest.

VmxExitInfo

VM-Exit Informations. (SDM Vol. 3C, Section 24.9.1)

VmxInterruptInfo

VM-Entry/VM-Exit Interruption-Information Field. (SDM Vol. 3C, Section 24.8.3, 24.9.2)

VmxIoExitInfo

Exit Qualification for I/O Instructions. (SDM Vol. 3C, Section 27.2.1, Table 27-5)

Enums

VmxExitReason

VMX basic exit reasons, as defined in the Intel Software Developer’s Manual (SDM) Vol. 3D, Appendix C.

Functions

has_hardware_support

Return if current platform support virtualization extension.